code review: core module

+++ src/terrama2/core/DataProviderDAO.cpp
@@ -27,82 +27,53 @@
   \author Paulo R. M. Oliveira
 */
 
+// TerraMA2
 #include "DataProviderDAO.hpp"
 #include "DataProvider.hpp"
 #include "DataSetDAO.hpp"
-#include "ApplicationController.hpp"
+#include "Exception.hpp"
 #include "Utils.hpp"
-#include "../Exception.hpp"
 
-// STL
-#include <vector>
-#include <memory>
-#include <cstdint>
-
-// terralib
-#include <terralib/common/StringUtils.h>
+// TerraLib
 #include <terralib/dataaccess/datasource/DataSourceTransactor.h>
-#include <terralib/memory.h>
+#include <terralib/memory/DataSet.h>
 #include <terralib/memory/DataSetItem.h>
 
-const std::string dataSetName = "terrama2.data_provider";
+// Qt
+#include <QObject>
 
-terrama2::core::DataProviderDAO::DataProviderDAO()
+static const std::string dataSetName = "terrama2.data_provider";
+
+void terrama2::core::DataProviderDAO::save(terrama2::core::DataProviderPtr dataProvider, te::da::DataSourceTransactor& transactor)
 {
-  std::shared_ptr<te::da::DataSource> dataSource = ApplicationController::getInstance().getDataSource();
-
-  if(!dataSource.get())
-  {
-    // PAULO-TODO: throw exception
-  }
-  transactor_ = dataSource->getTransactor();
-
-}
-
-terrama2::core::DataProviderDAO::~DataProviderDAO()
-{
-
-}
-
-
-void terrama2::core::DataProviderDAO::save(terrama2::core::DataProviderPtr dataProvider)
-{
-  if(!transactor_.get())
-  {
-    // PAULO-TODO: Throw exception to inform that the database connection is not available.
-  }
-
-  transactor_->begin();
-
-  // Removes the column id because it's an auto number
-  std::auto_ptr<te::da::DataSetType> dataSetType = transactor_->getDataSetType(dataSetName);
+// Removes the column id because it's an auto number
+  std::auto_ptr<te::da::DataSetType> dataSetType = transactor.getDataSetType(dataSetName);
   te::dt::Property* idProperty = dataSetType->getProperty(0);
   dataSetType->remove(idProperty);
 
-  // Creates a memory dataset from the DataSetType without column id
+// Creates a memory dataset from the DataSetType without column id
   std::shared_ptr<te::mem::DataSet> dataSet(new te::mem::DataSet(dataSetType.get()));
   te::mem::DataSetItem* dsItem = new te::mem::DataSetItem(dataSet.get());
 
-  // Sets the values in the item
+// Sets the values in the item
   dsItem->setString("name", dataProvider->name());
   dsItem->setString("description", dataProvider->description());
   dsItem->setInt32("kind", (int)dataProvider->kind());
   dsItem->setString("uri", dataProvider->uri());
   dsItem->setBool("active", DataProviderStatusToBool(dataProvider->status()));
 
-  // Adds it to the dataset
+// Adds it to the dataset
   dataSet->add(dsItem);
+
   std::map<std::string, std::string> options;
 
-  // Then, adds it to the data source
-  transactor_->add(dataSetName, dataSet.get(), options);
+// Then, adds it to the data source
+  transactor.add(dataSetName, dataSet.get(), options);
 
-  transactor_->commit();
-
-  // TODO: Remove this after getLastGeneratedId is implemented
+// TODO: Remove this after getLastGeneratedId is implemented
   std::string sql("SELECT id FROM " + dataSetName + " WHERE name = '" + dataProvider->name() + "'");
 
-  std::auto_ptr<te::da::DataSet> dataSetId = transactor_->query(sql);
+  std::auto_ptr<te::da::DataSet> dataSetId = transactor.query(sql);
 
   if(dataSetId->moveNext())
   {
@@ -110,82 +81,46 @@ void terrama2::core::DataProviderDAO::save(terrama2::core::DataProviderPtr dataP
   }
 
 
-  // Recovers the generated id and sets it in the provider
-  // TODO: Implement getLastGeneratedId in TerraLib
-  //dataProvider->setId(transactor_->getLastGeneratedId());
-
+// Recovers the generated id and sets it in the provider
+// TODO: Implement getLastGeneratedId in TerraLib
+//dataProvider->setId(transactor_->getLastGeneratedId());
 }
 
 
-void terrama2::core::DataProviderDAO::update(terrama2::core::DataProviderPtr dataProvider)
+void terrama2::core::DataProviderDAO::update(terrama2::core::DataProviderPtr dataProvider, te::da::DataSourceTransactor& transactor)
 {
-
-  if(!transactor_.get())
-  {
-    // PAULO-TODO: Throw exception to inform that the database connection is not available.
-  }
-
   if(dataProvider->id() == 0)
-  {
-    // PAULO-TODO: Throw exception to inform that the id is invalid.
-  }
-
-  transactor_->begin();
+    throw InvalidDataProviderIdError() << ErrorDescription(QObject::tr("Can not update a data provider with identifier: 0."));
 
   std::string sql = "UPDATE " + dataSetName + " SET"
       + " name='" + dataProvider->name() + "'"
       + ", description='" + dataProvider->description() + "'"
-      + ", kind=" + te::common::Convert2String((int)dataProvider->kind())
+      + ", kind=" + std::to_string(static_cast<int>(dataProvider->kind()))
       + ", uri='" + dataProvider->uri() + "'"
       + ", active=" + terrama2::core::BoolToString(DataProviderStatusToBool(dataProvider->status()))
-      + " WHERE id = " + te::common::Convert2String(dataProvider->id());
+  + " WHERE id = " + std::to_string(dataProvider->id());
 
-  transactor_->execute(sql);
-
-  transactor_->commit();
+  transactor.execute(sql);
 }
 
 
-void terrama2::core::DataProviderDAO::remove(DataProviderPtr dataProvider)
-{  
-  std::shared_ptr<te::da::DataSource> dataSource = ApplicationController::getInstance().getDataSource();
-  transactor_ = dataSource->getTransactor();
-  if(!transactor_.get())
-  {
-    // PAULO-TODO: Throw exception to inform that the database connection is not available.
-  }
-
-  transactor_->begin();
-
-  // Tries to removes all the datasets that belong to this provider
-
-  DataSetDAO datasetDAO;
-
-  auto dataSetList = dataProvider->dataSetList();
-  foreach (auto dataSet, dataSetList)
-  {
-    datasetDAO.remove(dataProvider->id(), transactor_);
-  }
+void terrama2::core::DataProviderDAO::remove(DataProviderPtr dataProvider, te::da::DataSourceTransactor& transactor)
+{
+  if(dataProvider->id() == 0)
+    throw InvalidDataProviderIdError() << ErrorDescription(QObject::tr("Can not remove a data provider with identifier: 0."));
 
   std::string sql = "DELETE FROM " + dataSetName
-      + " WHERE id = " + te::common::Convert2String(dataProvider->id());
+                  + " WHERE id = " + std::to_string(dataProvider->id());
 
-  transactor_->execute(sql);
-
-  transactor_->commit();
+  transactor.execute(sql);
 }
 
 
-terrama2::core::DataProviderPtr terrama2::core::DataProviderDAO::find(const uint64_t id) const
+terrama2::core::DataProviderPtr terrama2::core::DataProviderDAO::find(const uint64_t id, te::da::DataSourceTransactor& transactor)
 {
-  if(!transactor_.get())
-  {
-    // PAULO-TODO: Throw exception to inform that the database connection is not available.
-  }
+  std::string sql("SELECT * FROM " + dataSetName + " WHERE id = " + std::to_string(id));
 
-  std::string sql("SELECT * FROM " + dataSetName + " WHERE id = " + te::common::Convert2String(id));
-
-  std::auto_ptr<te::da::DataSet> dataSet = transactor_->query(sql);
+  std::auto_ptr<te::da::DataSet> dataSet = transactor.query(sql);
 
   DataProviderPtr provider;
 
@@ -204,17 +139,11 @@ terrama2::core::DataProviderPtr terrama2::core::DataProviderDAO::find(const uint
   return provider;
 }
 
-std::vector<terrama2::core::DataProviderPtr> terrama2::core::DataProviderDAO::list() const
+std::vector<terrama2::core::DataProviderPtr> terrama2::core::DataProviderDAO::list(te::da::DataSourceTransactor& transactor)
 {
-
-  if(!transactor_.get())
-  {
-    // PAULO-TODO: Throw exception to inform that the database connection is not available.
-  }
-
   std::vector<terrama2::core::DataProviderPtr> vecProviders;
 
-  std::auto_ptr<te::da::DataSet> dataSet = transactor_->getDataSet(dataSetName);
+  std::auto_ptr<te::da::DataSet> dataSet = transactor.getDataSet(dataSetName);
 
   while(dataSet->moveNext())
   {
